[IMP] GE03  made corrections on diferent files to improve general quality requirements and code
